+++ src/core/types.jl
@@ -35,12 +35,14 @@ SOCUBFModels = Union{SOCNLPUBFModel, SOCConicUBFModel}
 abstract type AbstractLPUBFModel <: AbstractNLPUBFModel end
 
 
-"Simplified BFM per Gan and Low 2014, PSCC, using matrix variables for power, voltage and current"
-abstract type LPfullUBFModel <: AbstractLPUBFModel end
-
-
 "LinDist3Flow per Sankur et al 2016, using vector variables for power, voltage and current"
-abstract type LPdiagUBFModel <: AbstractLPUBFModel end
+abstract type LPUBFDiagModel <: AbstractLPUBFModel end
+const LinDist3FlowModel = LPUBFDiagModel # more popular name for it
+
+
+"Simplified BFM per Gan and Low 2014, PSCC, using matrix variables for power, voltage and current"
+# This formulation is equivalent to LPUBFDiagModel, with redundant variables and constraints
+const LPUBFFullModel = LPUBFDiagModel
 
 
 "default SDP unbalanced DistFlow constructor"
@@ -60,16 +62,14 @@ mutable struct SOCConicUBFPowerModel <: SOCConicUBFModel _PMs.@pm_fields end
 
 
 "default LP unbalanced DistFlow constructor"
-mutable struct LPfullUBFPowerModel <: LPfullUBFModel _PMs.@pm_fields end
+mutable struct LPUBFDiagPowerModel <: LPUBFDiagModel _PMs.@pm_fields end
+const LinDist3FlowPowerModel = LPUBFDiagPowerModel # more popular name
 
 
-"default LP unbalanced DistFlow constructor"
-mutable struct LPdiagUBFPowerModel <: LPdiagUBFModel _PMs.@pm_fields end
+"Legacy name for LPUBFDiagModel; print a warning that this will be dropped soon."
+const LPUBFFullPowerModel = LPUBFDiagPowerModel
 
 
-"LinDist3Flow per Sankur et al 2016, using vector variables for power, voltage and current in scalar form"
-abstract type LPLinUBFModel <: _PMs.AbstractBFModel end
-
-
-"default Lin3Distflow constructor for scalar form"
-mutable struct LPLinUBFPowerModel <: LPLinUBFModel _PMs.@pm_fields end
+"This is completely equivalent to LPUBFDiagModel."
+#const LPLinUBFPowerModel = LPdiagUBFPowerModel
+mutable struct LPLinUBFPowerModel <: LPUBFDiagModel _PMs.@pm_fields end